requirements.txt
django==5.1
pillow==10.4.0
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
django-filter==24.3
requests==2.32.3
drf-yasg==1.21.7
drf-spectacular==0.27.2
django-cors-headers==4.4.0
celery
eventlet
redis
django-celery-beat
black==24.8.0
isort==5.13.2
